+++ ckanext/harvest/cli.py
@@ -108,8 +108,13 @@ def clear(ctx, id):
 
 @source.command()
 @click.argument(u"id", metavar=u"SOURCE_ID_OR_NAME", required=False)
+@click.option(
+    "-k",
+    "--keep-current",
+    default=False
+)
 @click.pass_context
-def clear_history(ctx, id):
+def clear_history(ctx, id, keep_current):
     """If no source id is given the history for all harvest sources
     (maximum is 1000) will be cleared.
 
@@ -122,7 +127,7 @@ def clear_history(ctx, id):
     flask_app = ctx.meta["flask_app"]
 
     with flask_app.test_request_context():
-        result = utils.clear_harvest_source_history(id)
+        result = utils.clear_harvest_source_history(id, bool(keep_current))
     click.secho(result, fg="green")
 
 
@@ -273,10 +278,10 @@ def run_test(ctx, id, force_import=None):
 
     """
     if force_import:
-        force_import_val = force_import.split('=')[-1]
+        force_import = force_import.split('=')[-1]
     flask_app = ctx.meta["flask_app"]
     with flask_app.test_request_context():
-        utils.run_test_harvester(id, force_import_val)
+        utils.run_test_harvester(id, force_import)
 
 
 @harvester.command("import")